.migration-backup/backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pydantic-settings>=2.4.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
python-dotenv>=1.0.1
supabase>=2.6.0
alembic>=1.13.0
httpx>=0.27.0
python-multipart>=0.0.9
requests>=2.31.0

redis>=5.0.0
pypdf>=4.0.0
ffmpeg-python>=0.2.0
cryptography>=42.0.0
